requirements.txt
elasticsearch==7.13.1
elasticsearch-dsl==7.4.0
pika==1.2.1
python-dateutil==2.8.2
PyYAML==5.4.1
git+https://github.com/cedadev/stac-generator
